wb_bot/requirements.txt
aiogram
apscheduler
python-dotenv
SQLAlchemy
